arch: extract business logic from MCP handlers (#195)
Move domain logic (reminder creation, inbox ops, graph execution,
task queue ops, state machine CRUD, agent lifecycle) from MCP
handler functions into a new `mcp_service` module. MCP handlers
now only parse JSON args and format JSON-RPC responses.
